+++ src/TsLinkPhp/IBatchCall.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Murdej\TsLinkPhp;
+
+interface IBatchCall
+{
+    public function batchCall(BatchCallEvent $event): void;
+}